rename framebuffer module
