+++ server/src/framebuffer/mod.rs
