+++ server/src/framebuffer/process.rs
