+++ server/src/framebuffer/reading.rs
